BS-11: moved database connection files into database folder and created postgres singleton
